category and product updated

+++ src/routers/categoryRouter.js
@@ -7,6 +7,7 @@ import {
 } from "../models/category/CategoryModel.js";
 import slugify from "slugify";
 import { updatCatValidation } from "../middlewares/joiMiddleware.js";
+import { getSelectedProduct } from "../model/products/productsModel.js";
 
 const router = express.Router();
 
@@ -87,6 +88,19 @@ router.put("/", updatCatValidation, async (res, req, next) => {
 // delete category
 router.delete("/:_id", async (req, res, next) => {
   const { _id } = req.params;
+
+  //get all the product that as parentCat === _id
+  const prodList = await getSelectedProduct({ parentCat: _id });
+  if (prodList.length) {
+    const names = prodList.map(({ name }) => name).toString();
+    return res.json({
+      status: "error",
+      message:
+        "Pelase re assign category for the following product before you can delete this category: " +
+        names,
+    });
+  }
+
   const result = await deleteCat(_id);
 
   if (result?._id) {
